+++ src/ReadoutEquipment.h
@@ -41,6 +41,10 @@ public:
   virtual void setDataOn();
   virtual void setDataOff();
 
+  // initialize / finalize counters (called before 1st loop and after last loop)
+  virtual void initCounters();
+  virtual void finalCounters();
+
   bool stopOnError = false; // if set, readout will stop when this equipment
                             // reports an error (isError flag)
   int isError = 0; // flag which might be used to count number of errors
@@ -49,6 +53,10 @@ public:
   // protected:
   // todo: give direct access to output FIFO?
   std::shared_ptr<AliceO2::Common::Fifo<DataBlockContainerReference>> dataOut;
+
+  // get current memory pool usage (available and total)
+  int getMemoryUsage(size_t &numberOfPagesAvailable,
+                     size_t &numberOfPagesInPool);
 
 private:
   std::unique_ptr<Thread> readoutThread;
@@ -91,7 +99,9 @@ protected:
     fifoOccupancyFreeBlocks = 10,
     fifoOccupancyReadyBlocks = 11,
     fifoOccupancyOutBlocks = 12,
-    maxIndex = 13 // not a counter, used to know number of elements in enum
+    nPagesUsed = 13, // number of used pages in memory pool
+    nPagesFree = 14, // number of free pages in memory pool
+    maxIndex = 15    // not a counter, used to know number of elements in enum
   };
 
   // Display names of the performance counters.
@@ -109,7 +119,9 @@ protected:
       "nPushedUp",
       "fifoOccupancyFreeBlocks",
       "fifoOccupancyReadyBlocks",
-      "fifoOccupancyOutBlocks"};
+      "fifoOccupancyOutBlocks",
+      "nPagesUsed",
+      "nPagesFree"};
 
   // check consistency (size) of EquipmentStatsNames with EquipmentStatsIndexes
   static_assert((sizeof(EquipmentStatsNames) /