+++ doc/releaseNotes.md
@@ -167,3 +167,9 @@ This file describes the main feature changes for each readout.exe released versi
   - default value for flushEquipmentTimeout changed to 1s.
 - equipment-player-*.autoChunk: fix metadata for FMQ-STF interface.
 - consumer-FairMQChannel-*: handling of RDH dynamic offset.
+
+## v1.3.0 - 22/01/2020
+- Updated the stopDma procedure for RORC device (new page reclaim mechanism).
+- Readout logs directed to stdout/stderr by default, when started from a console.
+- Added optional mode for interactive state machine (keyboard driven).
+- Fixes in the equipments for repetitive start/stop command cycles.